More test vectors
